minor uint//int warnings

+++ rollout.c
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: rollout.c,v 1.185 2008/01/04 18:40:52 Superfly_Jon Exp $
+ * $Id: rollout.c,v 1.186 2008/01/15 19:48:20 c_anthon Exp $
  */
 
 #include "config.h"
@@ -1682,7 +1682,7 @@ RolloutGeneral( ConstTanBoard *apBoard,
           aarMu[ alt ][ j ] = 0.0f;
       }
     } else {
-      int nGames = prc->nGamesDone;
+      uint nGames = prc->nGamesDone;
       double r;
 
       /* make sure the saved rollout contexts are consistent for
@@ -1694,7 +1694,7 @@ RolloutGeneral( ConstTanBoard *apBoard,
 	  prc->aecCubeLate[ i ].fCubeful = 
 	  prc->aecChequerLate[ i] .fCubeful = (prc->fCubeful || fCubeRollout);
 
-      if ((nGamesDone[ alt ] = nGames) < (int)nFirstTrial) 
+      if ((nGamesDone[ alt ] = nGames) < nFirstTrial) 
         nFirstTrial = nGames;
       /* restore internal variables from input values */
       for ( j = 0; j < NUM_ROLLOUT_OUTPUTS; ++j) {
